+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 3.0.0
 
+- Renames degit references to tiged ([#102](https://github.com/tiged/tiged/pull/102))
+
 - Breaking changes:
 
   - Change default export to a named export ([#101](https://github.com/tiged/tiged/pull/101))